Casting to double in call to log to avoid PGI compiler errors.
Former-commit-id: 52769d2fe0070ea418308e431993f7b23ca721da

+++ met/src/libcode/vx_statistics/contable_stats.cc
@@ -430,7 +430,8 @@ double TTContingencyTable::lodds() const {
       v = bad_data_double;
    }
    else {
-      v = log(fy_oy()) + log(fn_on()) - log(fy_on()) - log(fn_oy());
+      v = log((double) fy_oy()) + log((double) fn_on()) - 
+          log((double) fy_on()) - log((double) fn_oy());
    }
    
    return(v);
